requirements.txt
beautifulsoup4==4.6.0
boto==2.49.0
boto3==1.14.28
botocore==1.17.28
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
docutils==0.15.2
Flask==1.1.2
gensim==3.8.3
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lxml==4.5.2
MarkupSafe==1.1.1
numpy==1.19.1
feedparser-5.2.1
prawcore==1.4.0
python-dateutil==2.8.1
requests==2.24.0
s3transfer==0.3.3
scipy==1.5.2
six==1.15.0
smart-open==2.1.0
soupsieve==2.0.1
update-checker==0.17
urllib3==1.25.10
websocket-client==0.57.0
Werkzeug==1.0.1
